+++ http4k-client-apache/src/main/kotlin/org/http4k/client/ApacheClient.kt
@@ -4,17 +4,33 @@ import org.apache.http.Header
 import org.apache.http.StatusLine
 import org.apache.http.client.config.CookieSpecs.IGNORE_COOKIES
 import org.apache.http.client.config.RequestConfig
-import org.apache.http.client.methods.*
+import org.apache.http.client.methods.CloseableHttpResponse
+import org.apache.http.client.methods.HttpDelete
+import org.apache.http.client.methods.HttpEntityEnclosingRequestBase
+import org.apache.http.client.methods.HttpGet
+import org.apache.http.client.methods.HttpHead
+import org.apache.http.client.methods.HttpOptions
+import org.apache.http.client.methods.HttpRequestBase
+import org.apache.http.client.methods.HttpTrace
 import org.apache.http.conn.ConnectTimeoutException
 import org.apache.http.conn.HttpHostConnectException
 import org.apache.http.entity.ByteArrayEntity
 import org.apache.http.entity.InputStreamEntity
 import org.apache.http.impl.client.CloseableHttpClient
 import org.apache.http.impl.client.HttpClients
-import org.http4k.core.*
+import org.http4k.core.BodyMode
 import org.http4k.core.BodyMode.Memory
 import org.http4k.core.BodyMode.Stream
-import org.http4k.core.Method.*
+import org.http4k.core.Headers
+import org.http4k.core.HttpHandler
+import org.http4k.core.Method.DELETE
+import org.http4k.core.Method.GET
+import org.http4k.core.Method.HEAD
+import org.http4k.core.Method.OPTIONS
+import org.http4k.core.Method.TRACE
+import org.http4k.core.Request
+import org.http4k.core.Response
+import org.http4k.core.Status
 import org.http4k.core.Status.Companion.CLIENT_TIMEOUT
 import org.http4k.core.Status.Companion.CONNECTION_REFUSED
 import org.http4k.core.Status.Companion.UNKNOWN_HOST
@@ -32,13 +48,13 @@ object ApacheClient {
             try {
                 client.execute(request.toApacheRequest(requestBodyMode)).toHttp4kResponse(responseBodyMode)
             } catch (e: ConnectTimeoutException) {
-                Response(CLIENT_TIMEOUT.description("Client Error: caused by ${e.localizedMessage}"))
+                Response(CLIENT_TIMEOUT.toClientStatus(e))
             } catch (e: SocketTimeoutException) {
-                Response(CLIENT_TIMEOUT.description("Client Error: caused by ${e.localizedMessage}"))
+                Response(CLIENT_TIMEOUT.toClientStatus(e))
             } catch (e: HttpHostConnectException) {
-                Response(CONNECTION_REFUSED.description("Client Error: caused by ${e.localizedMessage}"))
+                Response(CONNECTION_REFUSED.toClientStatus(e))
             } catch (e: UnknownHostException) {
-                Response(UNKNOWN_HOST.description("Client Error: caused by ${e.localizedMessage}"))
+                Response(UNKNOWN_HOST.toClientStatus(e))
             }
         }