+++ http4k-core/src/main/kotlin/org/http4k/client/ext.kt
@@ -3,6 +3,7 @@ package org.http4k.client
 import org.http4k.core.HttpHandler
 import org.http4k.core.Request
 import org.http4k.core.Response
+import org.http4k.core.Status
 import java.io.Closeable
 
 interface AsyncHttpClient : Closeable {
@@ -20,3 +21,5 @@ fun HttpHandler.withAsyncApi(): AsyncHttpClient = object : DualSyncAsyncHttpHand
 
     override fun invoke(request: Request, fn: (Response) -> Unit) = fn(invoke(request))
 }
+
+fun Status.toClientStatus(e: Exception) = description("Client Error: $description caused by ${e.localizedMessage}")